+++ apps/wallet/src/ui/app/index.tsx
@@ -56,7 +56,11 @@ const HIDDEN_MENU_PATHS = [
 
 const App = () => {
 	const dispatch = useAppDispatch();
+<<<<<<< HEAD
 	const isSocialSignInEnabled = true; // useFeatureIsOn('enoki-social-sign-in');
+=======
+	const isSocialSignInEnabled = useFeatureIsOn('enoki-social-sign-in');
+>>>>>>> origin/main
 	const isPopup = useAppSelector((state) => state.app.appType === AppType.popup);
 	useEffect(() => {
 		document.body.classList.remove('app-initializing');